+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/pom.xml
@@ -125,7 +125,6 @@
                             version="${carbon.identity.package.import.version.range}",
                             javax.servlet.http; version="${imp.pkg.version.javax.servlet}"
                         </Import-Package>
-                        <DynamicImport-Package>*</DynamicImport-Package>
                         <Export-Package>
                             !org.wso2.carbon.identity.user.registration.mgt.internal,
                             org.wso2.carbon.identity.user.registration.mgt.*;
