+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/spotbugs-exclude.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+<FindBugsFilter>
+    <!--
+      There seems to be a bug in findbug, unable to read some bytecodes generated by java 11
+       properly, when try-with-resource pattern is used. This fails the following classes.
+       Please remove this exclusion when findbug plugin us updated from
+       org.codehaus.mojo:findbugs-maven-plugin-3.0.5 to a newer one which has the fix for the issue.
+       references,
+       https://github.com/spotbugs/spotbugs/issues/756
+    -->
+    <Match>
+        <Class name="org.wso2.carbon.identity.role.mgt.core.dao.GroupDAOImpl" />
+        <Bug pattern="RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE"/>
+    </Match>
+    <Match>
+        <Class name="org.wso2.carbon.identity.role.mgt.core.dao.RoleDAOImpl" />
+        <Bug pattern="RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE"/>
+    </Match>
+    <Match>
+        <Class name="org.wso2.carbon.identity.role.v2.mgt.core.dao.GroupDAOImpl" />
+        <Bug pattern="RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE"/>
+    </Match>
+    <Match>
+        <Class name="org.wso2.carbon.identity.role.v2.mgt.core.dao.RoleDAOImpl" />
+        <Bug pattern="RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE"/>
+    </Match>
+    <Match>
+        <Bug pattern="EI_EXPOSE_REP,EI_EXPOSE_REP2,MS_EXPOSE_REP"/>
+    </Match>
+    <Match>
+        <Bug pattern="SING_SINGLETON_GETTER_NOT_SYNCHRONIZED"/>
+    </Match>
+</FindBugsFilter>
