+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/src/main/java/org/wso2/carbon/identity/user/registration/mgt/dao/SQLConstants.java
@@ -25,31 +25,28 @@ public class SQLConstants {
 
     public static final String DELETE_FLOW =
             "DELETE FROM IDN_FLOW WHERE TENANT_ID = ? AND IS_DEFAULT = TRUE AND TYPE = ?";
-
     public static final String INSERT_FLOW_INTO_IDN_FLOW =
             "INSERT INTO IDN_FLOW (ID, TENANT_ID, FLOW_NAME, TYPE, IS_DEFAULT) VALUES (?, ?, ?, ?, ?)";
-
     public static final String INSERT_FLOW_NODE_INFO =
             "INSERT INTO IDN_FLOW_NODE (NODE_ID, FLOW_ID, NODE_TYPE, IS_FIRST_NODE) VALUES (?, ?, ?, ?)";
-
     public static final String INSERT_NODE_EXECUTOR_INFO =
             "INSERT INTO IDN_FLOW_NODE_EXECUTOR (FLOW_NODE_ID, EXECUTOR_NAME, IDP_NAME) VALUES (?, ?, ?)";
-
     public static final String INSERT_NODE_EDGES =
             "INSERT INTO IDN_FLOW_NODE_MAPPING (FLOW_NODE_ID, NEXT_NODE_ID, TRIGGERING_ELEMENT) VALUES (?, ?, ?)";
-
     public static final String INSERT_FLOW_PAGE_INFO =
             "INSERT INTO IDN_FLOW_PAGE (FLOW_ID, FLOW_NODE_ID, STEP_ID, PAGE_CONTENT, TYPE) VALUES (?, ?, ?, ?, ?)";
-
     public static final String INSERT_FLOW_PAGE_META =
             "INSERT INTO IDN_FLOW_PAGE_META (PAGE_ID, COORDINATE_X, COORDINATE_Y, HEIGHT, WIDTH) VALUES (?, ?, ?, ?, " +
                     "?)";
-
     public static final String GET_FLOW =
             "SELECT P.ID AS PAGE_ID, P.STEP_ID, P.PAGE_CONTENT, P.TYPE AS PAGE_TYPE, M.COORDINATE_X, M" +
                     ".COORDINATE_Y, M.HEIGHT, M.WIDTH FROM IDN_FLOW F JOIN IDN_FLOW_PAGE P ON F.ID = P.FLOW_ID " +
                     "LEFT JOIN IDN_FLOW_PAGE_META M ON P.ID = M.PAGE_ID WHERE F.TENANT_ID = ? AND F.IS_DEFAULT = " +
                     "TRUE AND F.TYPE = 'REGISTRATION';";
+
+    private SQLConstants() {
+
+    }
 
     /**
      * SQL Placeholders.
