+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/src/main/java/org/wso2/carbon/identity/user/registration/mgt/model/ComponentDTO.java
@@ -123,6 +123,9 @@ public class ComponentDTO implements Serializable {
         this.components = components;
     }
 
+    /**
+     * Builder class to build {@link ComponentDTO} objects.
+     */
     public static class Builder {
 
         private Map<String, Object> configs = new HashMap<>();
