+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/src/main/java/org/wso2/carbon/identity/user/registration/mgt/model/DataDTO.java
@@ -73,6 +73,9 @@ public class DataDTO implements Serializable {
         this.url = url;
     }
 
+    /**
+     * Builder class to build {@link DataDTO} objects.
+     */
     public static class Builder {
 
         private final List<ComponentDTO> components = new ArrayList<>();
