+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/src/main/java/org/wso2/carbon/identity/user/registration/mgt/model/ExecutorDTO.java
@@ -70,6 +70,9 @@ public class ExecutorDTO implements Serializable {
         this.idpName = idpName;
     }
 
+    /**
+     * Builder class to build {@link ExecutorDTO}.
+     */
     public static class Builder {
 
         private String name;
