+++ components/user-self-registration/org.wso2.carbon.identity.user.registration.mgt/src/main/java/org/wso2/carbon/identity/user/registration/mgt/utils/RegistrationMgtUtils.java
@@ -25,6 +25,10 @@ import org.wso2.carbon.identity.user.registration.mgt.exception.RegistrationServ
 
 public class RegistrationMgtUtils {
 
+    private RegistrationMgtUtils() {
+
+    }
+
     /**
      * Handle the registration flow management server exceptions.
      *
